Intermediate step for PointValue Type cleanup ... next is get Mango Value out of PointValueTime ...

+++ scadabr/scadabr/src/main/java/com/serotonin/mango/vo/dataSource/http/HttpRetrieverPointLocatorVO.java
@@ -66,7 +66,7 @@ public class HttpRetrieverPointLocatorVO extends AbstractPointLocatorVO {
                 }
             }
 
-            if (vo.dataType == DataType.NUMERIC && !vo.valueFormat.isEmpty()) {
+            if (vo.dataType == DataType.DOUBLE && !vo.valueFormat.isEmpty()) {
                 try {
                     //TODO Localization with server locale !!!
                     new DecimalFormat(vo.valueFormat);
